move containerFilter to own filter.go file
